fix(wework): avoid empty message source rows
